Move misc stuff to utils module
